# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Router",
    "version": "generated"
  },
  "paths": {
...
    "/api/v1/route": {
      "post": {
        "tags": [
          "routing"
        ],
        "summary": "Route",
        "operationId": "route_api_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/principals": {
      "get": {
        "tags": [
          "policy"
        ],
        "summary": "List Principals",
        "operationId": "list_principals_api_v1_principals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrincipalOut"
                  },
                  "type": "array",
                  "title": "Response List Principals Api V1 Principals Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "policy"
        ],
        "summary": "Create Principal",
        "operationId": "create_principal_api_v1_principals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrincipalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalCreated"
                }
              }
            }
          },
          "409": {
            "description": "agentId exists, or dev mode without MCPR_ADMIN_TOKEN"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/principals/{principal_id}": {
      "get": {
        "tags": [
          "policy"
        ],
        "summary": "Get Principal By Id",
        "operationId": "get_principal_by_id_api_v1_principals__principal_id__get",
        "parameters": [
          {
            "name": "principal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Principal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "policy"
        ],
        "summary": "Patch Principal",
        "operationId": "patch_principal_api_v1_principals__principal_id__patch",
        "parameters": [
          {
            "name": "principal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Principal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrincipalPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "policy"
        ],
        "summary": "Delete Principal",
        "operationId": "delete_principal_api_v1_principals__principal_id__delete",
        "parameters": [
          {
            "name": "principal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Principal Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/principals/{principal_id}/rotate-key": {
      "post": {
        "tags": [
          "policy"
        ],
        "summary": "Rotate Key",
        "operationId": "rotate_key_api_v1_principals__principal_id__rotate_key_post",
        "parameters": [
          {
            "name": "principal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Principal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyRotated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policy-rules": {
      "get": {
        "tags": [
          "policy"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_policy_rules_get",
        "parameters": [
          {
            "name": "agentId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agentid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleOut"
                  },
                  "title": "Response List Rules Api V1 Policy Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "policy"
        ],
        "summary": "Create Rule",
        "operationId": "create_rule_api_v1_policy_rules_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policy-rules/{rule_id}": {
      "patch": {
        "tags": [
          "policy"
        ],
        "summary": "Patch Rule",
        "operationId": "patch_rule_api_v1_policy_rules__rule_id__patch",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "policy"
        ],
        "summary": "Delete Rule",
        "operationId": "delete_rule_api_v1_policy_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals": {
      "get": {
        "tags": [
          "policy"
        ],
        "summary": "List Approvals",
        "description": "Newest first, at most `limit` (default 200, the old silent cap). The\nresponse stays a bare list; an unknown `status` is 422, not `[]`.",
        "operationId": "list_approvals_api_v1_approvals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "executing",
                    "executed",
                    "failed",
                    "denied",
                    "expired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalOut"
                  },
                  "title": "Response List Approvals Api V1 Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/approve": {
      "post": {
        "tags": [
          "policy"
        ],
        "summary": "Approve",
        "operationId": "approve_api_v1_approvals__approval_id__approve_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/deny": {
      "post": {
        "tags": [
          "policy"
        ],
        "summary": "Deny",
        "operationId": "deny_api_v1_approvals__approval_id__deny_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/approvals/{approval_id}": {
      "get": {
        "tags": [
          "policy"
        ],
        "summary": "My Approval",
        "operationId": "my_approval_api_v1_me_approvals__approval_id__get",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyRotated": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agentId": {
            "type": "string",
            "title": "Agentid"
          },
          "apiKey": {
            "type": "string",
            "title": "Apikey"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agentId",
          "apiKey"
        ],
        "title": "KeyRotated"
      },
...
      "RouteBody": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 4000,
            "title": "Query"
          },
          "agentId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentid"
          },
          "maxTools": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxtools"
          },
          "maxServers": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxservers"
          },
          "allowedServers": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 120,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowedservers"
          },
          "maxSkills": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxskills"
          },
          "kinds": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "tool",
                    "skill"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Kinds"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RouteBody"
      },
...
    }
  }
}